Improve game state namespacing
