+++ src/main/kotlin/engine/gameState/GameContactListener.kt
@@ -1,4 +1,4 @@
-package engine
+package engine.gameState
 
 import engine.freeBody.Warhead
 import org.jbox2d.callbacks.ContactImpulse
